requirements.txt
GitPython
alive_progress
angr
angrop
apscheduler
autopep8
bandit
binwalk
boofuzz
capstone
crypto
cstruct
cython
debugpy
frida-tools
ipdb
jefferson
keystone-engine
libheap
log_symbols
maat
nuitka
objection
openai
opencv-python
osquery
pandas
pebble
pefile
poetry
progress
pwncli
pwnscripts
pwntools
pyarmor
pycipher
pycryptodome
pyradamsa
python-magic
qiling
r2pipe
randcrack
randomname
rfcat
ropgadget
ropper
rsatool
sagemath
scapy
scikit-learn
semantic-version 
smmap2
spinners
stego-lsb
sudo
ubi_reader
uncompyle6
unicorn
urwid
xortool
z3-solver
zeek
